+++ src/TreeHouse/QueueBundle/DependencyInjection/TreeHouseQueueExtension.php
@@ -20,6 +20,36 @@ use TreeHouse\Queue\Processor\Retry\RetryProcessor;
 class TreeHouseQueueExtension extends Extension
 {
     /**
+     * @var string[]
+     */
+    private $connections = [];
+
+    /**
+     * @var string[]
+     */
+    private $exchanges = [];
+
+    /**
+     * @var string[]
+     */
+    private $queues = [];
+
+    /**
+     * @var string[]
+     */
+    private $publishers = [];
+
+    /**
+     * @var string[]
+     */
+    private $consumers = [];
+
+    /**
+     * @var array Map that links exchanges with a DLX counterpart
+     */
+    private $dlxs = [];
+
+    /**
      * @inheritdoc
      */
     public function load(array $configs, ContainerBuilder $container)
@@ -36,6 +66,8 @@ class TreeHouseQueueExtension extends Extension
         $this->loadConsumers($config, $container);
         $this->loadExchanges($config, $container);
         $this->loadQueues($config, $container);
+
+        $this->setCreatedDefinitionsParameters($container);
 
         if (!$config['auto_flush']) {
             $container->removeDefinition('tree_house.queue.event_listener.queue');
@@ -98,14 +130,9 @@ class TreeHouseQueueExtension extends Extension
      */
     private function loadPublishers(array $config, ContainerBuilder $container)
     {
-        $publishers = [];
         foreach ($config['publishers'] as $name => $publisher) {
-            $publisherId = $this->createPublisherDefinition($name, $publisher, $container);
-            $publishers[$name] = $publisherId;
+            $this->createPublisherDefinition($name, $publisher, $container);
         }
-
-        // set a parameter to reference the publishers
-        $container->setParameter('tree_house.queue.publishers', $publishers);
     }
 
     /**
@@ -114,15 +141,9 @@ class TreeHouseQueueExtension extends Extension
      */
     private function loadConsumers(array $config, ContainerBuilder $container)
     {
-        $consumers = [];
-
         foreach ($config['consumers'] as $name => $consumer) {
-            $consumerId = $this->createConsumerDefinition($name, $consumer, $container);
-            $consumers[$name] = $consumerId;
+            $this->createConsumerDefinition($name, $consumer, $container);
         }
-
-        // set a parameter to reference the consumers
-        $container->setParameter('tree_house.queue.consumers', $consumers);
     }
 
     /**
@@ -131,8 +152,8 @@ class TreeHouseQueueExtension extends Extension
      */
     private function loadExchanges(array $config, ContainerBuilder $container)
     {
-        foreach ($config['exchanges'] as $name => $queue) {
-            $this->createExchangeDefinition($name, $queue, $container);
+        foreach ($config['exchanges'] as $name => $exchange) {
+            $this->createExchangeDefinition($name, $exchange, $container);
         }
     }
 
@@ -179,6 +200,8 @@ class TreeHouseQueueExtension extends Extension
         $channelId = sprintf('tree_house.queue.channel.%s', $name);
         $container->setDefinition($channelId, $definition);
 
+        $this->connections[$name] = $connectionId;
+
         return $connectionId;
     }
 
@@ -196,6 +219,7 @@ class TreeHouseQueueExtension extends Extension
         $connection = $config['connection'] ?: $container->getParameter('tree_house.queue.default_connection');
         $channelId = sprintf('tree_house.queue.channel.%s', $connection);
         $channelAlias = sprintf('tree_house.queue.channel.%s', $name);
+        $exchangeName = $config['name'] ?: $name;
 
         // add alias if connection is named differently than exchange
         if ($name !== $connection) {
@@ -206,13 +230,26 @@ class TreeHouseQueueExtension extends Extension
         $definition = new Definition($container->getParameter('tree_house.queue.exchange.class'));
         $definition->setFactory([$amqpFactory, 'createExchange']);
         $definition->addArgument(new Reference($channelAlias));
-        $definition->addArgument($name);
+        $definition->addArgument($exchangeName);
         $definition->addArgument($config['type']);
         $definition->addArgument($this->getExchangeFlagsValue($config));
         $definition->addArgument($config['arguments']);
 
         $exchangeId = sprintf('tree_house.queue.exchange.%s', $name);
         $container->setDefinition($exchangeId, $definition);
+
+        $this->exchanges[$name] = $exchangeId;
+
+        if (isset($config['dlx']['enable']) && $config['dlx']['enable']) {
+            if (!isset($config['dlx']['name'])) {
+                $config['dlx']['name'] = sprintf('%s.dead', $exchangeName);
+            }
+
+            $dlxName = $config['dlx']['name'];
+            $dlxId = $this->createExchangeDefinition($dlxName, $config['dlx'], $container);
+
+            $this->dlxs[$name] = $dlxId;
+        }
 
         return $exchangeId;
     }
@@ -246,6 +283,8 @@ class TreeHouseQueueExtension extends Extension
         $consumerId = sprintf('tree_house.queue.consumer.%s', $name);
         $container->setDefinition($consumerId, $definition);
 
+        $this->consumers[$name] = $consumerId;
+
         return $consumerId;
     }
 
@@ -262,6 +301,13 @@ class TreeHouseQueueExtension extends Extension
 
         $connection = $queue['connection'] ?: $container->getParameter('tree_house.queue.default_connection');
         $channelId = sprintf('tree_house.queue.channel.%s', $connection);
+        $arguments = $queue['arguments'];
+
+        // if there is an exchange with the same name, and it has a DLX configured, set this in the arguments
+        if (!array_key_exists('x-dead-letter-exchange', $arguments) && $dlxId = $this->getDeadLetterExchange($name)) {
+            $dlx = $container->getDefinition($dlxId);
+            $arguments['x-dead-letter-exchange'] = $dlx->getArgument(1);
+        }
 
         // create queue
         $definition = new Definition($container->getParameter('tree_house.queue.queue.class'));
@@ -269,7 +315,7 @@ class TreeHouseQueueExtension extends Extension
         $definition->addArgument(new Reference($channelId));
         $definition->addArgument($queue['name']);
         $definition->addArgument($this->getQueueFlagsValue($queue));
-        $definition->addArgument($queue['arguments']);
+        $definition->addArgument($arguments);
 
         if (empty($queue['bindings'])) {
             // bind to the same exchange
@@ -293,6 +339,8 @@ class TreeHouseQueueExtension extends Extension
 
         $queueId = sprintf('tree_house.queue.queue.%s', $name);
         $container->setDefinition($queueId, $definition);
+
+        $this->queues[$name] = $queueId;
 
         return $queueId;
     }
@@ -320,6 +368,8 @@ class TreeHouseQueueExtension extends Extension
         $publisher->addArgument(new Reference($composerId));
 
         $container->setDefinition($publisherId, $publisher);
+
+        $this->publishers[$name] = $publisherId;
 
         return $publisherId;
     }
@@ -465,6 +515,18 @@ class TreeHouseQueueExtension extends Extension
     }
 
     /**
+     * @param ContainerBuilder $container
+     */
+    private function setCreatedDefinitionsParameters(ContainerBuilder $container)
+    {
+        $container->setParameter('tree_house.queue.connections', $this->connections);
+        $container->setParameter('tree_house.queue.exchanges', $this->exchanges);
+        $container->setParameter('tree_house.queue.queues', $this->queues);
+        $container->setParameter('tree_house.queue.publishers', $this->publishers);
+        $container->setParameter('tree_house.queue.consumers', $this->consumers);
+    }
+
+    /**
      * @param array $exchange
      *
      * @return int
@@ -510,5 +572,19 @@ class TreeHouseQueueExtension extends Extension
         }
 
         return $flags;
+    }
+
+    /**
+     * @param string $name
+     *
+     * @return string|null
+     */
+    private function getDeadLetterExchange($name)
+    {
+        if (isset($this->dlxs[$name])) {
+            return $this->dlxs[$name];
+        }
+
+        return null;
     }
 }